requirement1.txt
pandas
streamlit
streamlit_pandas_profiling
sys
os
torch
transformer
openpyxl
xlrd